Added glows
When quote is retrieved, glow will be added based on the house of the relevent character.
If the house they belong to is on screen, their house section will glow (drop shadow), if it is out of scroll view, bottom edge of window will glow

Need to fix todo on the HouseSection component where I filter twice.
Need to clean up typing where 'X is possibly null'.
Need to style fetch button.
Need to focus on other stretch objectives